requirements.txt
numpy
pandas
scikit-learn
tensorflow
matplotlib
seaborn
jupyter
ipykernel
pyserial